Coltrol de excepciones
Agregando excepciones personalizadas

+++ src/main/java/com/sventas/sventas/controller/ClienteController.java
@@ -1,5 +1,6 @@
 package com.sventas.sventas.controller;
 
+import com.sventas.sventas.exception.ModelNotFoundException;
 import com.sventas.sventas.model.Cliente;
 import com.sventas.sventas.service.ClienteService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -33,12 +34,20 @@ public class ClienteController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> findByid(@PathVariable("id") Integer idCliente){
-        return new ResponseEntity<>(clienteService.findById(idCliente), HttpStatus.OK) ;
+    public ResponseEntity<Cliente> findById(@PathVariable("id") Integer idCliente){
+        Cliente cliente = clienteService.findById(idCliente);
+        if (cliente==null){
+            throw new ModelNotFoundException("Cliente no encontrado");
+        }
+        return new ResponseEntity<>(cliente, HttpStatus.OK) ;
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> delete(@PathVariable("id") Integer idCliente){
+    public ResponseEntity<Object> delete(@PathVariable("id") Integer idCliente) throws Exception {
+        Cliente cliente = clienteService.findById(idCliente);
+        if (cliente==null){
+            throw new ModelNotFoundException("El cliente no existe!");
+        }
         clienteService.delete(idCliente);
         return new ResponseEntity<>(HttpStatus.OK);
     }
